Add files via upload
Tic-Tac-Toe game in MIPS Assembly Language.
This program lets you play a simulation of Tic-Tac-Toe. Is program with validation and the normal rules in which you can find in a real game of Tic-Tac-Toe. The program also come with sound. Every turn will come with a sound and every time someone wins, it will play a small jingle.
Purpose: To learn the new advance language of MIPS and to discover what you can build with it.
How to Run: Download MARS (An IDE for the mips language) then download the file. Open the game within MARS, compile then run. You now can play a game of Tic-Tac-Toe with anyone.
Unit Width in Pixels : 1
Unit Height in Pixels: 1
Display Width in Pixels: 256
Display Height in Pixels: 256